==> 1_transcript_audio_batch.ipynb <==
import pandas as pd

import torch
from transformers import AutoModelForSpeechSeq2Seq, AutoProcessor, pipeline
from datasets import load_dataset


device = "cuda:0" if torch.cuda.is_available() else "cpu"
torch_dtype = torch.float16 if torch.cuda.is_available() else torch.float32

# model_id = "openai/whisper-large-v3"
# model_id = "openai/whisper-small"
model_id = "openai/whisper-medium"
# model_id = "openai/whisper-tiny"

model = AutoModelForSpeechSeq2Seq.from_pretrained(
    model_id, torch_dtype=torch_dtype, low_cpu_mem_usage=True, use_safetensors=True
)
model.to(device)

processor = AutoProcessor.from_pretrained(model_id)

pipe = pipeline(
    "automatic-speech-recognition",
    model=model,
    tokenizer=processor.tokenizer,
    feature_extractor=processor.feature_extractor,
    max_new_tokens=128,
    chunk_length_s=30,
    batch_size=16,
    return_timestamps=True,
    torch_dtype=torch_dtype,
    device=device,
)


import os
basepath = 'data/english_course_wma/Wizard Inglês Modulo 3 - Conversation'

_, _, filenames = next( os.walk(basepath) )
filenames

filenames = sorted( filenames )

for filename in filenames[5:]:
    data = []
    print( filename )
    complete_path = os.path.join(basepath, filename)

    result = pipe(complete_path, generate_kwargs={"language": "english"})

    for i, row in enumerate( result['chunks'] ):
        # print( row['text'] )

        data.append( row["text"] )

    df = pd.read_csv("transcript.csv")

    joined = " ".join(data)
    row = { 
        "Filename": filename,
        "Sentences": joined
        }

    df = pd.concat( [df, pd.DataFrame([row])], ignore_index=True )

    # print(df)
    df.to_csv( "transcript.csv", index=False )

# df.to_csv( "transcript.csv" )

df

print( df.iloc[1,2] )

print( df.iloc[2,2] )
