==> 2_separate_senteces_batch.ipynb <==
# from langchain_google_genai import ChatGoogleGenerativeAI

# # llm = ChatGoogleGenerativeAI(model="gemini-pro", convert_system_message_to_human=True, google_api_key="AIzaSyC-V6lfROehy46ntB6zPZ7CJ8zNF3gDdO4")
# llm = ChatGoogleGenerativeAI(model="gemini-1.5-pro-latest", convert_system_message_to_human=True, google_api_key="AIzaSyC-V6lfROehy46ntB6zPZ7CJ8zNF3gDdO4")

# from langchain_community.llms import Ollama
# llm = Ollama(
#     model="llama2:7b", 
#     base_url = 'http://host.docker.internal:11434'
#     )

from g4f import Provider, models
from langchain.llms.base import LLM

from langchain_g4f import G4FLLM

llm: LLM = G4FLLM(
    model=models.gpt_35_turbo,
    # provider=Provider.Aichat,
)

from langchain.prompts import PromptTemplate

prompt = PromptTemplate.from_template(
    """
    I'm studying english. I have a list of senteces that I took from a English Book.
    I want you to put the string "-separator-" to separate related senteces.
    To clarify what are related, I give an example that contains the before and after process.
    You should do similiar to the NEW DATA TO PROCESS.

    Also, each sample conversation starts with "Conversation, Lesson #" with the number of the lesson. 
    you should output it and put a "-separator-" right after it to separate it from the sentences.

    You should answer only with the new senteces, you shold not output the original text or something like this.

    # EXAMPLE BEFORE PROCESS:
    Music  Wizard Book 3 Conversation Lesson 1  Did the boy pay $1 for the paper at the newsstand?  Yes, he did. He paid $1 for the paper at the newsstand.  Did this secretary put the document on the table?  Yes, she did.  She put the document on the table.  No, she didn't.  She didn't put the information from the computer?  Yes, I did. I accessed the information from the computer.  No, I didn't. I didn't access the information from the computer.  Did your mom do her Christmas shopping early this year?  Yes, she did.  She did her Christmas shopping early this year.  No, she didn't. She didn't do her Christmas shopping early this year.  Did they eat Brazilian food at the restaurant? Yes, they did.  They ate Brazilian food at the restaurant.  No, they didn't.  They didn't eat Brazilian food at the restaurant. Is your sister doing her homework in the her bedroom?  Yes, she is.  She is doing her homework in her bedroom.  No, she isn't. She isn't doing her homework in her bedroom.  Are the children playing in the snow?  Yes, they are. They are playing in the snow.  Now they aren't. They aren't playing in the snow.  Is your son putting the toys in the room?  Yes he is.  He is putting the toys in the room.  No, he isn't. He isn't putting the toys in the room.  Is Jane accessing her email every day?  Yes she is. She is accessing her email every day. No, she isn't. She isn't accessing her email every day.  Is the man paying for the groceries?  Yes he is.  He is paying for the groceries.  No he isn't.  He is not paying for the groceries.
    
    # EXAMPLE AFTER PROCESS:

    WIZARD Book 3 Conversation
    Lesson 1
    -separator-
    Did the boy pay one dollar for the paper at the newsstand?
    Yes, he did. He paid one dollar for the paper at the newsstand.
    No, he didn't. He didn't pay one dollar for the paper at the newsstand. 
    -separator-
    Did the secretary put the document on the table.
    No, she didn't. She didn't put the document on the table.
    -separator-
    Did you access the information from the computer?
    Yes, I did. I access the information from the computer.
    -separator-
    Did your mom do her Christmas shopping early this year?
    Yes, she did. She did her Christmas shopping early this year.
    No, she didn't. She didn't do her Christmas shopping early this year.
    -separator-
    Did they eat Brazilian food at the restaurant.
    No, they didn't. They didn't eat Brazilian food at the restaurant.
    -separator-
    Is your sister doing her homework in her bedroom?
    Yes, she is. She is doing her homework in her bedroom.
    No, she isn't. She isn't doing her homework in her bedroom.
    -separator-
    Are the children playing in the snow?
    Yes, they are. They are playing in the snow.
    No, they aren't. They aren't playing in the snow.
    -separator-
    Is your son putting the toys in the room?
    Yes, he is. He is putting the toys in the room.
    No, he isn't. He isn't putting the toys in the room.
    -separator-
    Is Jane accessing her email every day?
    Yes, she is. She is accessing her email every day. 
    No, she isn't. She isn't accessing her email every day.
    -separator-
    Is the man paying for the groceries?
    Yes he is. He is paying for the groceries.
    No he isn't. He is not paying for the groceries.



    # NEW DATA TO PROCESS:

    {new_data_to_process}

    """
)

chain = prompt | llm

import pandas as pd

df = pd.read_csv("transcript.csv", index_col=None )
df

data = []
for i, row in df.iterrows():
    new_data_to_process = row["Sentences"]

    response = chain.invoke({"new_data_to_process": new_data_to_process})
    
    # print( response.content )
    # data.append( response.content )
    data.append( response ) #if llama2

    # if i==1:
    #     break


data

df_separated = df[["Filename"]]
df_separated

df_separated["Sentences_separated"] = data
df_separated

# print(df_separated.iloc[1,2])

df_separated.to_csv( "transcript_separated.csv", index=False )
