==> 3_create_new_sentences_batch.ipynb <==
# from langchain_google_genai import ChatGoogleGenerativeAI

# # llm = ChatGoogleGenerativeAI(model="gemini-pro", convert_system_message_to_human=True, google_api_key="AIzaSyC-V6lfROehy46ntB6zPZ7CJ8zNF3gDdO4")
# # llm = ChatGoogleGenerativeAI(model="gemini-1.5-pro-latest", convert_system_message_to_human=True, google_api_key="AIzaSyC-V6lfROehy46ntB6zPZ7CJ8zNF3gDdO4")
# llm = ChatGoogleGenerativeAI(model="gemini-1.5-pro-latest", convert_system_message_to_human=True, google_api_key="AIzaSyClqzPkQR8HlCfvztCYTuv35hmNxRYqZ58")

# from langchain_community.llms import Ollama
# llm = Ollama(
#     model="llama2:7b", 
#     base_url = 'http://host.docker.internal:11434'
#     )

from g4f import Provider, models
from langchain.llms.base import LLM

from langchain_g4f import G4FLLM

llm: LLM = G4FLLM(
    model=models.gpt_35_turbo,
    # provider=Provider.Aichat,
)

from langchain.prompts import PromptTemplate

prompt = PromptTemplate.from_template(
    """
    I'm studying english to get a machine learning job in a IT company. 
    I will provide you with a list of sentences from an english book. 
    I need you to change each sentence to a similiar sentence, but using vocabulary from another range of domain like work, meetings, machine learning, development, programming.
    You could use words to persons like developer, programmer, etc.

    Also, each sample conversation starts with "Conversation, Lesson #" with the number of the lesson. 
    You should repeat the title and you should put a "-separator-" after it to separate it from the sentences.
    You should mantain the overall structyure of data to process.

    Abstain of creating explanatory pieace of text like: "Here is your answer as you asked, new english senteces etc...
    You should not print "# NEW DATA TO PROCESS:" as this is just a piece of text to explain where the text to process start. 
    The content you output will be used in a programatic way.

    You should answer only with the convesation title and the new senteces, you shold not output the original text or something like this.

    # NEW DATA TO PROCESS:

    {new_data_to_process}

    """
)

chain = prompt | llm

import pandas as pd

df = pd.read_csv("transcript_separated.csv")
df

data = []

for i, row in df.iloc[:].iterrows():
    print(i, row["Filename"])
    # print( row["Sentences"] )

    new_data_to_process = row["Sentences_separated"]

    response = chain.invoke({"new_data_to_process": new_data_to_process})

    data.append( response )
    
    # if i==0:
    #     break

data

df_new = df[["Filename"]]
df_new

data[1:]

df_new["Sentences_new"] = data[1:]

df_new

df_new.to_csv( "transcript_new_sentences.csv", index=False )
