==> 4_generate_text_to_speech.ipynb <==
import pandas as pd

from transformers import SpeechT5Processor, SpeechT5ForTextToSpeech, SpeechT5HifiGan
from datasets import load_dataset
import torch
import soundfile as sf
from datasets import load_dataset

processor = SpeechT5Processor.from_pretrained("microsoft/speecht5_tts")
model = SpeechT5ForTextToSpeech.from_pretrained("microsoft/speecht5_tts")
vocoder = SpeechT5HifiGan.from_pretrained("microsoft/speecht5_hifigan")

# load xvector containing speaker's voice characteristics from a dataset
embeddings_dataset = load_dataset("Matthijs/cmu-arctic-xvectors", split="validation")
speaker_embeddings = torch.tensor(embeddings_dataset[7306]["xvector"]).unsqueeze(0)

df = pd.read_csv("transcript_new_sentences.csv")
df

import os

for i, row in df.iterrows():
    lesson_name = row["Filename"].split(".")[0]
    print( i, lesson_name )

    folder = f"new_audio/{lesson_name}"
    if not os.path.exists(folder):
        os.makedirs(folder)

    for j, line in enumerate( row["Sentences_new"].replace("-separator-", "").split("\n")[1:] ):
        if line != "":
            print( line )
            

            inputs = processor(text=line, return_tensors="pt")
            speech = model.generate_speech(inputs["input_ids"], speaker_embeddings, vocoder=vocoder)

            sf.write(f"{folder}/{j:03d}.wav", speech.numpy(), samplerate=16000)

    #     if j >0:
    #         break

    # if i == 0:
    #     break
