==> 5_join_audio_files.ipynb <==
import soundfile as sf
import numpy as np
import os

def generate_silence(duration, samplerate):
    num_samples = int(duration * samplerate)
    return np.zeros((num_samples,))

for i, (root, dir, files) in enumerate( list( os.walk("new_audio") )[1:] ):
    filename = root.split("/")[1]
    files = sorted( files )
    
    print( i, filename , files )

    # Initialize an empty list to store concatenated audio data
    concatenated_data = []

    # Loop through each audio file
    for i, file_path in enumerate(files):
        # Load audio data from the current file
        data, samplerate = sf.read(f"{root}/{file_path}")
        
        # Append the audio data to the concatenated_data list
        concatenated_data.append(data)
        
        # add silence of the same durantion of the last audio
        silence_duration = len(data) / samplerate
        silence_data = generate_silence(silence_duration, samplerate)
        concatenated_data.append(silence_data)

    # Concatenate all audio data into a single array
    final_data = np.concatenate(concatenated_data)

    # Specify the output file path where you want to save the concatenated audio
    output_file = f'new_audio_joined/{filename}.wav'

    # Write the concatenated audio data to the specified output file
    sf.write(output_file, final_data, samplerate)

    # if i >= 0:
    #     break
