==> 6_convert_to_mp3.ipynb <==
import os
from pydub import AudioSegment


i = 0
for root, dirs, files in os.walk("new_audio"):
    if i == 0:
        i += 1
        continue

    files = [ f for f in files if f.split(".")[1] == "wav" ]
    print( root, dirs, files )

    for file in files:
        new_filename = f"{root}/{file.split('.')[0]}"
        wav_file = AudioSegment.from_wav(f"{new_filename}.wav")

        # Convert WAV to MP3
        wav_file.export(f"{new_filename}.mp3", format="mp3")
    
    #     if i == 1:
    #         break

    # if i == 1:
    #     break



i = 0
for root, dirs, files in os.walk("new_audio_joined"):
    if i == 0:
        i += 1
        continue

    files = [ f for f in files if f.split(".")[1] == "wav" ]
    print( root, dirs, files )

    for file in files:
        new_filename = f"{root}/{file.split('.')[0]}"
        os.remove(f"{new_filename}.wav")

    
    #     if i == 1:
    #         break

    # if i == 1:
    #     break
